python/art_sales/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
boto3==1.40.47
botocore==1.40.47
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
croniter==6.0.0
dependency-injector==4.48.2
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
rq==2.6.0
s3transfer==0.14.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3